Change Bogon Status -> Scope Filter
See comments in migration 39e4f15ed80c for explanation of these unusual
migrations. Also sets alembic to use a transaction per migration.

+++ irrd/storage/alembic/versions/39e4f15ed80c_add_bogon_status.py
@@ -7,26 +7,40 @@ Create Date: 2020-04-22 14:43:57.985437
 """
 from alembic import op
 import sqlalchemy as sa
+from sqlalchemy.engine.reflection import Inspector
 
 
 # revision identifiers, used by Alembic.
+
 revision = '39e4f15ed80c'
 down_revision = '1743f98a456d'
 branch_labels = None
 depends_on = None
 
 
-def upgrade():
-    bogon_status = sa.Enum('unknown', 'not_bogon', 'bogon_as', 'bogon_prefix', name='bogonstatus')
-    bogon_status.create(op.get_bind())
+# This is a very strange migration.
+# The bogon_status column was added at some point, but was never used.
+# Some people's deployments have this migration already deployed, so
+# it can't be removed. For anyone upgrading from earlier, we don't
+# want to re-add it to then remove it in the next migration, because
+# it takes a lot of time.
+#
+# Therefore, this migration does nothing to upgrade, and on downgrade
+# removes the column, but only if it exists.
+# It assumes the existence of the bogonstatus enum matches that of
+# the bogon_status column.
 
-    op.add_column('rpsl_objects', sa.Column('bogon_status', sa.Enum('unknown', 'not_bogon', 'bogon_as', 'bogon_prefix', name='bogonstatus'), nullable=False, server_default='unknown'))
-    op.create_index(op.f('ix_rpsl_objects_bogon_status'), 'rpsl_objects', ['bogon_status'], unique=False)
+def upgrade():
+    pass
 
 
 def downgrade():
-    op.drop_index(op.f('ix_rpsl_objects_bogon_status'), table_name='rpsl_objects')
-    op.drop_column('rpsl_objects', 'bogon_status')
+    conn = op.get_bind()
+    inspector = Inspector.from_engine(conn)
+    columns = inspector.get_columns('rpsl_objects')
 
-    bogon_status = sa.Enum('unknown', 'not_bogon', 'bogon_as', 'bogon_prefix', name='bogonstatus')
-    bogon_status.drop(op.get_bind())
+    if 'bogon_status' in columns:
+        op.drop_index(op.f('ix_rpsl_objects_bogon_status'), table_name='rpsl_objects')
+        op.drop_column('rpsl_objects', 'bogon_status')
+        bogon_status = sa.Enum('unknown', 'not_bogon', 'bogon_as', 'bogon_prefix', name='bogonstatus')
+        bogon_status.drop(op.get_bind())
